backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.135.3",
    "openai-whisper>=20250625",
    "pytest>=9.0.3",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.44.0",
]
